Added args to converters
This will be used in more complex conversions that need to know some details about how to convert. This is mainly added to prevent having to implement a dozen different converts of the same type to handle a few minor changes.

Examples of usage will include arrays, maps, sets, and constructors.