Added support for lists
Bit more than just lists as it supports any collection both for save and load. However, this is specifically set as list since the output from load is an ArrayList. Meaning for method injection it has to be a list to work. 

For field injection it will support any collection and even a consumer function. This way we actually have a much broader way of handling data at the moment. In the future I will add specific handling for common data structures to allow more ways of doing methods.